refactor(medicion_sensor): change value of var

+++ lib/features/map/domain/entities/medicion_sensor.dart
@@ -1,6 +1,6 @@
 class MedicionSensor {
   final int idContenedor;
-  final int volumenMedido;
+  final double volumenMedido;
   final String nivelLlenado;
 
   const MedicionSensor({
